Refactor directory structure for differnet exceptions
