+++ src/sexpr/cast.cpp
@@ -1,4 +1,4 @@
-#include "../../include/repl/EvalException.hpp"
+#include "../../include/repl/except/EvalException.hpp"
 #include <memory>
 #include <sstream>
